# api_service/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for managing purchase transactions with multi-currency support.",
        "title": "Purchase Transaction API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/transactions": {
            "post": {
                "description": "Validates and stores a transaction payload, then publishes it for asynchronous processing",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "transactions"
                ],
                "summary": "Create a new purchase transaction",
                "parameters": [
                    {
                        "description": "Transaction Payload",
                        "name": "transaction",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/ports.TransactionRequestDTO"
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Accepted",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/transactions/{id}/convert": {
            "get": {
                "description": "Fetches the result of a conversion job from Valkey",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "transactions"
                ],
                "summary": "Retrieve a conversion result",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Transaction ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Target Currency Code (e.g. BRL)",
                        "name": "currency",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ports.TransactionResponseDTO"
                        }
                    },
                    "404": {
                        "description": "Result not found yet",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "post": {
                "description": "Triggers an asynchronous conversion job for the specified currency",
                "tags": [
                    "transactions"
                ],
                "summary": "Request a currency conversion",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Transaction ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Target Currency Code (e.g. BRL)",
                        "name": "currency",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Accepted",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Invalid Request",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/transactions/{id}/status": {
            "get": {
                "description": "Retrieves the current status from Valkey",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "transactions"
                ],
                "summary": "Check transaction processing status",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Transaction ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Status",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid UUID",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Status not found",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "ports.TransactionRequestDTO": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "purchase_amount": {
                    "type": "number"
                },
                "transaction_date": {
                    "description": "YYYY-MM-DD",
                    "type": "string"
                }
            }
        },
        "ports.TransactionResponseDTO": {
            "type": "object",
            "properties": {
                "convertedAmount": {
                    "type": "number"
                },
                "description": {
                    "type": "string"
                },
                "exchangeRate": {
                    "type": "number"
                },
                "id": {
                    "type": "string"
                },
                "purchaseAmountUSD": {
                    "type": "number"
                },
                "targetCurrency": {
                    "type": "string"
                },
                "transactionDate": {
                    "type": "string"
                }
            }
        }
    }
}